+++ polars/polars-lazy/Cargo.toml
@@ -79,6 +79,7 @@ meta = []
 pivot = ["polars-core/rows", "polars-ops/pivot"]
 top_k = ["polars-ops/top_k"]
 semi_anti_join = ["polars-core/semi_anti_join"]
+cse = []
 
 # no guarantees whatsoever
 private = ["polars-time/private"]
