+++ py-polars/Cargo.toml
@@ -84,6 +84,7 @@ all = [
   # we need to add this, as maturin fails if we don't
   # remove this once polars-algo is released
   "polars/algo",
+  "polars/cse",
 ]
 
 # we cannot conditionaly activate simd
